+++ src/features/epigram-create/model/useEpigramCreate.ts
@@ -1,13 +1,11 @@
-import { useState } from "react";
-
-import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
 import { createEpigram } from "@/entities/epigram/api/createEpigram";
+import { useTagInput } from "@/shared/hooks/useTagInput";
 
 import { AUTHOR_TYPE, type EpigramCreateFormValues } from "./schema";
-
-const MAX_TAG_LENGTH = 10;
 
 function resolveAuthor(values: EpigramCreateFormValues, userNickname: string | undefined): string {
   if (values.authorType === AUTHOR_TYPE.UNKNOWN) return "알 수 없음";
@@ -18,7 +16,7 @@ function resolveAuthor(values: EpigramCreateFormValues, userNickname: string | u
 export function useEpigramCreate(userNickname?: string) {
   const router = useRouter();
   const queryClient = useQueryClient();
-  const [tagInput, setTagInput] = useState("");
+  const { tagInput, handleTagInputChange, handleAddTag, handleRemoveTag } = useTagInput();
 
   const mutation = useMutation({
     mutationFn: (values: EpigramCreateFormValues) =>
@@ -34,25 +32,6 @@ export function useEpigramCreate(userNickname?: string) {
       router.push(`/epigrams/${epigram.id}`);
     },
   });
-
-  function handleTagInputChange(value: string): void {
-    setTagInput(value.slice(0, MAX_TAG_LENGTH));
-  }
-
-  function handleAddTag(currentTags: string[], onChange: (tags: string[]) => void): void {
-    const trimmed = tagInput.trim();
-    if (!trimmed || currentTags.length >= 3 || currentTags.includes(trimmed)) return;
-    onChange([...currentTags, trimmed]);
-    setTagInput("");
-  }
-
-  function handleRemoveTag(
-    tag: string,
-    currentTags: string[],
-    onChange: (tags: string[]) => void
-  ): void {
-    onChange(currentTags.filter((t) => t !== tag));
-  }
 
   return {
     tagInput,
